+++ src/pid/ServoAnalogPIDMotor.h
@@ -31,7 +31,7 @@ public:
 	void attach(int servoPin, int analogPin);
 	int64_t getPosition();
 
-	void setOutput(int64_t out);
+	void setOutput(int32_t out);
 	void overrideCurrentPositionHardware(int64_t val);
 	double calcCur(void);
 };
